# openapi3-v1.yaml
openapi: 3.0.1
info:
  title: Job Manager Service - API v1
  description: Job Manager Service API version 1
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
...
  /v1/jobs/{jobId}/user-metadata:
    patch:
      operationId: updateUserMetadata
      parameters:
        - $ref: '#/components/parameters/jobId'
      summary: Update job's custom metadata
      description: >
        Updates the user-defined metadata object for a specific job.

        This endpoint allows clients to attach or modify arbitrary data related
        to a job

        without affecting the job's core properties or execution status.


        User metadata is useful for storing application-specific context,
        tracking information,

        or any custom data needed by client applications.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userMetadata'
      responses:
        '200':
          description: User metadata successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid metadata format or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
...
  /v1/jobs/{jobId}/status:
    parameters:
      - $ref: '#/components/parameters/jobId'
    put:
      operationId: updateStatus
      summary: Update job status for workflow control
      description: >
        Updates the operational status of a job to control its workflow
        execution. This endpoint 

        supports user-initiated operations including PENDING (resume), ABORTED
        (cancel), and 

        PAUSED (suspend), which may cascade changes to related stages and tasks.


        Internal status transitions (such as IN_PROGRESS, COMPLETED, FAILED,
        CREATED) are managed 

        automatically by the system based on stage and task completion, and
        cannot be set through 

        this endpoint.


        Status changes follow a state machine that enforces valid transitions.
        When a job's status 

        is changed, the system will automatically update timestamps and
        completion percentages as 

        appropriate.
      tags:
        - jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/jobOperationStatus'
              required:
                - status
      responses:
        '200':
          description: Job status successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: JOB_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid status or illegal state transition
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - ILLEGAL_JOB_STATUS_TRANSITION
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
...
  /v1/stages/{stageId}/tasks:
    parameters:
      - $ref: '#/components/parameters/stageId'
    get:
      operationId: getTasksByStageId
      summary: Retrieve all tasks for a specific stage
      description: >
        Fetches all tasks associated with the specified stage ID.

        Provides complete information about each task including type, status,
        and attempt count.
      responses:
        '200':
          description: Successfully retrieved tasks for the specified stage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/taskResponse'
        '400':
          description: Invalid stage ID format or other parameter error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
      tags:
        - stages
    post:
      operationId: addTasks
      summary: Add new tasks to an existing stage
      description: >
        Creates and appends new tasks to an existing stage.

        This endpoint allows for extending stage processing capabilities by
        adding more work units.


        Task objects require type and data properties, with optional user
        metadata and 

        maximum attempt configuration. Tasks are created with an initial status
        of PENDING.


        The stage must exist and be in a valid state to accept new tasks.
      parameters:
        - $ref: '#/components/parameters/stageId'
      tags:
        - stages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/createTaskPayload'
            examples:
              file_processing_tasks:
                summary: Multiple file processing tasks
                description: Batch creation of tasks for processing multiple files
                value:
                  - data:
                      file_path: /data/input/file_001.csv
                      output_path: /data/processed/file_001.parquet
                      transformation: csv_to_parquet
                    userMetadata:
                      file_size: 125MB
                      estimated_time: 5m
                    maxAttempts: 3
                  - data:
                      file_path: /data/input/file_002.csv
                      output_path: /data/processed/file_002.parquet
                      transformation: csv_to_parquet
                    userMetadata:
                      file_size: 89MB
                      estimated_time: 3m
                    maxAttempts: 3
                  - data:
                      file_path: /data/input/file_003.csv
                      output_path: /data/processed/file_003.parquet
                      transformation: csv_to_parquet
                    userMetadata:
                      file_size: 256MB
                      estimated_time: 8m
                    maxAttempts: 5
              api_integration_tasks:
                summary: API integration tasks with different configurations
                value:
                  - data:
                      endpoint: https://api.external.com/customers
                      method: GET
                      headers:
                        Authorization: Bearer ${API_TOKEN}
                      pagination:
                        page_size: 100
                        max_pages: 50
                    userMetadata:
                      source: customer_system
                      priority: high
                  - data:
                      endpoint: https://api.external.com/orders
                      method: GET
                      headers:
                        Authorization: Bearer ${API_TOKEN}
                      query_params:
                        start_date: '2025-01-01'
                        end_date: '2025-07-27'
                    userMetadata:
                      source: order_system
                      priority: medium
              tasks_with_tracing:
                summary: Tasks with distributed tracing context
                description: Task creation with tracing information for observability
                value:
                  - data:
                      process_type: data_enrichment
                      input_dataset: customer_profiles
                      enrichment_sources:
                        - demographic_api
                        - credit_api
                    traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
                    tracestate: customer=enrich123,priority=high
                    maxAttempts: 2
                  - data:
                      process_type: data_validation
                      validation_rules:
                        - email_format
                        - phone_format
                    traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-11a268bb4ba902b7-01
                    maxAttempts: 1
      responses:
        '201':
          description: Tasks successfully created and added to the stage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/taskResponse'
              examples:
                file_processing_tasks_response:
                  summary: Response for multiple file processing tasks creation
                  value:
                    - id: 3f09f17d-029e-417f-b84e-d6782efc2371
                      data:
                        file_path: /data/input/file_001.csv
                        output_path: /data/processed/file_001.parquet
                        transformation: csv_to_parquet
                      stageId: 5251e4cb-3f34-4be7-812f-dca3fa6e733a
                      userMetadata:
                        file_size: 125MB
                        estimated_time: 5m
                      status: PENDING
                      attempts: 0
                      maxAttempts: 3
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:30:00Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-55e6abbf8fd902b7-01
                      tracestate: file=001,processing=csv_parquet
                    - id: 469e8a0c-9232-4144-920f-fb5cef356ff7
                      data:
                        file_path: /data/input/file_002.csv
                        output_path: /data/processed/file_002.parquet
                        transformation: csv_to_parquet
                      stageId: 6e1e2e25-22ee-4e24-b394-c544d253ff43
                      userMetadata:
                        file_size: 89MB
                        estimated_time: 3m
                      status: PENDING
                      attempts: 0
                      maxAttempts: 3
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:30:00Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-66f7bcc090e902b7-01
                      tracestate: file=002,processing=csv_parquet
                    - id: 6a7ecf04-4455-41d1-aa04-8786178cb4a3
                      data:
                        file_path: /data/input/file_003.csv
                        output_path: /data/processed/file_003.parquet
                        transformation: csv_to_parquet
                      stageId: 714ef37c-218f-4818-a166-cbc13d2850a6
                      userMetadata:
                        file_size: 256MB
                        estimated_time: 8m
                      status: PENDING
                      attempts: 0
                      maxAttempts: 5
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:30:00Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-77081dd1a1f902b7-01
                      tracestate: file=003,processing=csv_parquet
                api_integration_tasks_response:
                  summary: Response for API integration tasks creation
                  value:
                    - id: ddfad658-da33-4573-b5a3-d70a204a3e0b
                      data:
                        endpoint: https://api.external.com/customers
                        method: GET
                        headers:
                          Authorization: Bearer ${API_TOKEN}
                        pagination:
                          page_size: 100
                          max_pages: 50
                      stageId: a09866f1-74f1-49c1-b996-268586ad346d
                      userMetadata:
                        source: customer_system
                        priority: high
                      status: PENDING
                      attempts: 0
                      maxAttempts: 3
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:30:00Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-88192ee2b20902b7-01
                      tracestate: api=customers,source=external
                    - id: df59d74e-f442-4e66-b8c9-b312ad948c47
                      data:
                        endpoint: https://api.external.com/orders
                        method: GET
                        headers:
                          Authorization: Bearer ${API_TOKEN}
                        query_params:
                          start_date: '2025-01-01'
                          end_date: '2025-07-27'
                      stageId: 4daa6a5c-da33-4cbe-8e4a-dbc00182fcfa
                      userMetadata:
                        source: order_system
                        priority: medium
                      status: PENDING
                      attempts: 0
                      maxAttempts: 3
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:30:00Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-99203ff3c31902b7-01
                      tracestate: api=orders,source=external
                tasks_with_lifecycle_states:
                  summary: >-
                    Tasks showing different lifecycle states with timing
                    information
                  description: >-
                    Examples of tasks in various states showing startTime and
                    endTime usage
                  value:
                    - id: f1a2b3c4-5678-90ab-cdef-123456789012
                      data:
                        process_type: data_processing
                        input_file: /data/raw/dataset_001.json
                      stageId: 1a2b3c4d-5e6f-7890-ab12-345678901234
                      status: PENDING
                      attempts: 0
                      maxAttempts: 3
                      userMetadata: {}
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:30:00Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-aa111bb222c902b7-01
                      tracestate: state=pending
                    - id: f2a3b4c5-6789-01bc-def2-345678901234
                      data:
                        process_type: data_validation
                        input_file: /data/raw/dataset_002.json
                      stageId: 2b3c4d5e-6f78-9012-bc34-567890123456
                      status: IN_PROGRESS
                      attempts: 1
                      maxAttempts: 3
                      userMetadata: {}
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:32:15Z'
                      startTime: '2025-07-27T10:32:15Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-bb222cc333d902b7-01
                      tracestate: state=in_progress
                    - id: f3a4b5c6-789a-12cd-ef34-567890123456
                      data:
                        process_type: data_export
                        output_file: /data/processed/dataset_003.parquet
                      stageId: 3c4d5e6f-7890-1234-cd56-789012345678
                      status: COMPLETED
                      attempts: 1
                      maxAttempts: 3
                      userMetadata: {}
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:35:45Z'
                      startTime: '2025-07-27T10:33:20Z'
                      endTime: '2025-07-27T10:35:45Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-cc333dd444e902b7-01
                      tracestate: state=completed
                    - id: f4a5b6c7-89ab-23de-f456-789012345678
                      data:
                        process_type: data_cleanup
                        cleanup_rules:
                          - remove_duplicates
                          - validate_schema
                      stageId: 4d5e6f78-9012-3456-de78-901234567890
                      status: FAILED
                      attempts: 3
                      maxAttempts: 3
                      userMetadata: {}
                      creationTime: '2025-07-27T10:30:00Z'
                      updateTime: '2025-07-27T10:38:12Z'
                      startTime: '2025-07-27T10:36:05Z'
                      endTime: '2025-07-27T10:38:12Z'
                      traceparent: 00-4bf92f3577b34da6a3ce929d0e0e4736-dd444ee555f902b7-01
                      tracestate: state=failed
        '400':
          description: Invalid request format or validation error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for bad request errors, providing
                      a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - VALIDATION_ERROR
                          - STAGE_IN_FINITE_STATE
                          - NOT_ALLOWED_TO_ADD_TASKS_TO_IN_PROGRESS_STAGE
        '404':
          description: Stage not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stageNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
  /v1/stages/{stageType}/tasks/dequeue:
    parameters:
      - $ref: '#/components/parameters/stageType'
    patch:
      operationId: dequeueTask
      summary: Find and claim the highest priority pending task of specified stage type
      description: >
        Retrieves the highest priority task of the specified stage type that is
        in PENDING or RETRIED status, 

        and automatically updates its status to IN_PROGRESS. This endpoint
        implements a priority-based 

        work queue pattern where workers can claim the next available task.


        The endpoint considers task priority (inherited from the parent job),
        searches only for tasks 

        that are in valid states (PENDING or RETRIED), and updates related stage
        and job status if needed.


        If successful, returns the complete task details with status updated to
        IN_PROGRESS.
      tags:
        - stages
      responses:
        '200':
          description: Task successfully dequeued and status updated to IN_PROGRESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskResponse'
        '400':
          description: Invalid stageType parameter or other validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: No pending tasks of requested type are available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskNotFoundResponse'
        '500':
          description: Internal server error or invalid state transition
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/baseErrorResponse'
                  - type: object
                    description: >
                      Error response structure for internal server errors,
                      providing a consistent

                      format for common-readable error codes and human-readable
                      messages.
                    properties:
                      code:
                        enum:
                          - DATABASE_RELATED_ERROR
                          - UNKNOWN_ERROR
                          - TASK_STATUS_UPDATE_FAILED
                          - ILLEGAL_TASK_STATUS_TRANSITION
                          - ILLEGAL_STAGE_STATUS_TRANSITION
                          - ILLEGAL_JOB_STATUS_TRANSITION
                          - STAGE_NOT_FOUND
                          - JOB_NOT_FOUND
  /v1/tasks:
    parameters:
      - $ref: '#/components/parameters/paramStageId'
      - $ref: '#/components/parameters/paramStageType'
      - $ref: '#/components/parameters/fromDate'
      - $ref: '#/components/parameters/endDate'
      - $ref: '#/components/parameters/paramsTaskStatus'
    get:
      operationId: getTasksByCriteria
      summary: Retrieve tasks matching specified criteria
      description: >
        Returns a filtered list of tasks based on the provided query parameters.

        Supports filtering by stage ID, task type, date range, and operational
        status.


        This endpoint is useful for monitoring task progress across multiple
        stages and jobs,

        enabling clients to build custom dashboards or track specific task
        types.
      responses:
        '200':
          description: Successfully retrieved matching tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/taskResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
      tags:
        - tasks
  /v1/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/taskId'
    get:
      operationId: getTaskById
      summary: Retrieve a specific task by its ID
      description: >
        Fetches detailed information about a task using its unique identifier.

        Returns complete task data including type, status, payload, and attempt
        information.
      responses:
        '200':
          description: Task data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskResponse'
        '400':
          description: Invalid task ID format or other parameter error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
      tags:
        - tasks
  /v1/tasks/{taskId}/user-metadata:
    patch:
      operationId: updateTaskUserMetadata
      parameters:
        - $ref: '#/components/parameters/taskId'
      summary: Update task's custom metadata
      description: >
        Updates the user-defined metadata object for a specific task.

        This endpoint allows clients to attach or modify arbitrary data related
        to a task

        without affecting the task's core properties or execution status.


        User metadata is useful for storing application-specific context,
        tracking information,

        or any custom data needed by client applications.
      tags:
        - tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userMetadata'
      responses:
        '200':
          description: User metadata successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/defaultOkMessage'
              example:
                code: TASK_MODIFIED_SUCCESSFULLY
        '400':
          description: Invalid metadata format or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/badRequestResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/taskNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internalErrorsResponse'
...
security: []
components:
  parameters:
    jobId:
      in: path
      name: jobId
      description: Unique identifier for the job
      required: true
      schema:
        $ref: '#/components/schemas/jobId'
    stageId:
      in: path
      name: stageId
      description: Unique identifier for the stage
      required: true
      schema:
        $ref: '#/components/schemas/stageId'
...
  schemas:
...
    jobId:
      type: string
      format: uuid
      description: Unique identifier for a job
...
    jobOperationStatus:
      type: string
      enum:
        - PENDING
        - ABORTED
        - PAUSED
      example: PENDING
      description: >
        User-controllable job statuses: PENDING (resume), ABORTED (cancel),
        PAUSED (suspend).

        System-managed: IN_PROGRESS, COMPLETED, FAILED, CREATED.
...
    userMetadata:
      type: object
      additionalProperties: true
      default: {}
      description: >-
        Application-specific custom data container that can store arbitrary
        client information without affecting core operations
...
    defaultOkMessage:
      type: object
      description: |
        Standard success response structure used for operations that don't
        return entity data, providing a standardized confirmation message.
      properties:
        code:
          $ref: '#/components/schemas/successMessages'
      required:
        - code
    baseErrorResponse:
      type: object
      description: >
        Foundation schema for standardized error responses across all API
        endpoints.

        Provides consistent structure with human-readable message and
        common-readable code fields

        that are extended by individual endpoints to define their specific error
        scenarios.
      properties:
        message:
          type: string
          description: Human-readable error message describing the issue
        code:
          type: string
          description: common-readable error code indicating the specific error type
      required:
        - message
        - code
    internalErrorsResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Provides a unified error response format for internal server errors,
            including a common-readable error code and a clear, human-readable
            message to help clients identify and handle issues consistently.
          properties:
            code:
              enum:
                - DATABASE_RELATED_ERROR
                - UNKNOWN_ERROR
          required:
            - code
    badRequestResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response structure for bad request errors,
            providing a consistent

            format for common-readable error codes and human-readable messages.
          properties:
            code:
              enum:
                - VALIDATION_ERROR
          required:
            - code
    jobNotFoundResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response for cases where a requested job is not
            found in the system.

            Provides a consistent format for common-readable error codes and
            human-readable messages.
          properties:
            code:
              enum:
                - JOB_NOT_FOUND
          required:
            - code
    stageNotFoundResponse:
      allOf:
        - $ref: '#/components/schemas/baseErrorResponse'
        - type: object
          description: >
            Standardized error response for cases where a requested stage is not
            found in the system.

            Provides a consistent format for common-readable error codes and
            human-readable messages.
          properties:
            code:
              enum:
                - STAGE_NOT_FOUND
          required:
            - code
...