# src/api/v1/tasks/controller.ts
import type { Logger } from '@map-colonies/js-logger';
import httpStatus from 'http-status-codes';
import { injectable, inject } from 'tsyringe';
import { HttpError } from '@map-colonies/error-express-handler';
import type { TypedRequestHandlers } from '@openapi';
import { SERVICES, successMessages } from '@common/constants';
import {
  IllegalJobStatusTransitionError,
  IllegalStageStatusTransitionError,
  IllegalTaskStatusTransitionError,
  JobNotFoundError,
  NotAllowedToAddTasksToInProgressStageError,
  StageInFiniteStateError,
  StageNotFoundError,
  TaskStatusUpdateFailedError,
  TaskNotFoundError,
} from '@src/common/generated/errors';
import { TaskManager } from '../../../tasks/models/manager';
import { type TasksFindCriteriaArg } from '../../../tasks/models/models';

const badRequestErrors = [
  TaskStatusUpdateFailedError,
  IllegalTaskStatusTransitionError,
  IllegalStageStatusTransitionError,
  IllegalJobStatusTransitionError,
  JobNotFoundError,
  StageNotFoundError,
];

const internalErrors = [
  TaskStatusUpdateFailedError,
  IllegalTaskStatusTransitionError,
  IllegalStageStatusTransitionError,
  IllegalJobStatusTransitionError,
  JobNotFoundError,
  StageNotFoundError,
];

@injectable()
export class TaskControllerV1 {
  public constructor(
    @inject(SERVICES.LOGGER) private readonly logger: Logger,
    @inject(TaskManager) private readonly manager: TaskManager
  ) {}

  public addTasks: TypedRequestHandlers['POST /v1/stages/{stageId}/tasks'] = async (req, res, next) => {
    try {
      const response = await this.manager.addTasks(req.params.stageId, req.body);

      return res.status(httpStatus.CREATED).json(response);
    } catch (err) {
      if (err instanceof StageNotFoundError) {
        (err as HttpError).status = httpStatus.NOT_FOUND;
      }

      if (err instanceof NotAllowedToAddTasksToInProgressStageError || err instanceof StageInFiniteStateError) {
        (err as HttpError).status = httpStatus.BAD_REQUEST;
      }

      return next(err);
    }
  };

  public getTasks: TypedRequestHandlers['GET /v1/tasks'] = async (req, res, next) => {
    const params: TasksFindCriteriaArg = req.query;
    try {
      const response = await this.manager.getTasks(params);
      return res.status(httpStatus.OK).json(response);
    } catch (err) {
      this.logger.error(`Error occurred on getting tasks with error`, err);
      next(err);
    }
  };

  public getTaskById: TypedRequestHandlers['GET /v1/tasks/{taskId}'] = async (req, res, next) => {
    try {
      const response = await this.manager.getTaskById(req.params.taskId);
      return res.status(httpStatus.OK).json(response);
    } catch (err) {
      if (err instanceof TaskNotFoundError) {
        (err as HttpError).status = httpStatus.NOT_FOUND;
      }

      return next(err);
    }
  };

  public getTaskByStageId: TypedRequestHandlers['GET /v1/stages/{stageId}/tasks'] = async (req, res, next) => {
    try {
      const response = await this.manager.getTasksByStageId(req.params.stageId);
      return res.status(httpStatus.OK).json(response);
    } catch (err) {
      if (err instanceof StageNotFoundError) {
        (err as HttpError).status = httpStatus.NOT_FOUND;
      }
      next(err);
    }
  };

  public updateUserMetadata: TypedRequestHandlers['PATCH /v1/tasks/{taskId}/user-metadata'] = async (req, res, next) => {
    try {
      await this.manager.updateUserMetadata(req.params.taskId, req.body);
      return res.status(httpStatus.OK).json({ code: successMessages.taskModifiedSuccessfully });
    } catch (err) {
      if (err instanceof TaskNotFoundError) {
        (err as HttpError).status = httpStatus.NOT_FOUND;
      }

      return next(err);
    }
  };

  public updateStatus: TypedRequestHandlers['PUT /v1/tasks/{taskId}/status'] = async (req, res, next) => {
    try {
      const response = await this.manager.updateStatus(req.params.taskId, req.body.status);

      return res.status(httpStatus.OK).json(response);
    } catch (err) {
      if (err instanceof TaskNotFoundError) {
        (err as HttpError).status = httpStatus.NOT_FOUND;
      } else if (badRequestErrors.some((e) => err instanceof e)) {
        (err as HttpError).status = httpStatus.BAD_REQUEST;
        this.logger.error({ msg: `Task status update failed: invalid status transition`, status: req.body.status, err });
      }

      return next(err);
    }
  };

  public dequeue: TypedRequestHandlers['PATCH /v1/stages/{stageType}/tasks/dequeue'] = async (req, res, next) => {
    try {
      const response = await this.manager.dequeue(req.params.stageType);

      return res.status(httpStatus.OK).json(response);
    } catch (err) {
      if (err instanceof TaskNotFoundError) {
        (err as HttpError).status = httpStatus.NOT_FOUND;
      } else if (internalErrors.some((e) => err instanceof e)) {
        (err as HttpError).status = httpStatus.INTERNAL_SERVER_ERROR;
      }

      return next(err);
    }
  };
}